Only target paired breakout exits during fanout
